requirements.txt
psycopg2==2.8.3
requests==2.22.0
Flask_Cors==3.0.8
numpy==1.17.2
Flask_WTF==0.14.2
Flask==1.1.1
Flask_Dance==2.2.0
pandas==0.25.1
WTForms==2.2.1
six==1.12.0
pymemcache==2.2.2
pandas==0.25.1
board==1.0
pyserial==3.4
